feature: add standard named method implementations with @method_implementation support
* new MulleObjCStandardImplementation module providing canonical implementations for +alloc, +new, -init, -autorelease, -dealloc, -finalize, -release, -retain, -retainCount and -self
* new @method_implementation directive lets classes opt into standard implementations by name instead of writing inline bodies
* fast-path dispatch (e.g. +instance) now calls inline standard implementations via a class implementation mask for zero overhead
* standard implementations are single-sourced: the public named functions delegate to the same inline core, so custom overrides and the fast path stay in sync

Refactored NSObject (+instance, -init, -self) and MulleObjCRootObject (+alloc, +new, -release, -retain, -retainCount, -autorelease, -dealloc, -finalize) to use the new standard implementations.

fix: correct NSThread universe-membership teardown order

* split _mulle_objc_thread_resignas_universethread so config/TSS, global ABA and universe-release steps peel in strict LIFO order, preventing a NULL-deref/SIGSEGV when universe teardown runs on the releasing thread

Added tests covering @method_implementation (attribute inheritance, plain objects) and the thread acquire/relinquish failure path.

+++ test/NSInvocation/metaabi.m
@@ -19,7 +19,7 @@ extern void  *__forward_mulle_objc_object_call( id, SEL, ...);
 
 - (int) intReturnNoParam                       { return( 1848); }
 - (int) intReturnIntParam:(int) param          { return( param + 1); }
-- (int) intReturnPointerParam:(void *) param   { return( (int) param + 1); }
+- (int) intReturnPointerParam:(void *) param   { return( (int) (intptr_t) param + 1); }
 - (int) intReturnDoubleParam:(double) param    { return( (int) param + 1); }
 - (int) intReturnStructParam:(NSRange) param   { return( (int) ((param.location + 1) * 100 + param.length + 1)); }
